% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df1_env}
\alias{df1_env}
\title{Simulated environmental factor distribution}
\format{
A data frame with 100 rows and 2 columns:
\describe{
  \item{patch}{Habitat patch identifier.}
  \item{env}{Environmental factor level.}
}
}
\source{
A \code{FilterABM} simulation using the following code (random seed was not used so the output may vary).

\code{lh = init_envt(env_mean_lh = 0, env_sd_lh = 3, npatch = 100, gradient = "correlated", rho = 0.95)}

\code{df1_env <- lh \%>\% distinct(patch, env)}
}
\usage{
df1_env
}
\description{
The results of a simulation of environmental filtering using the \code{FilterABM} package.
The environment consists of 100 habitat patches where the environmental factor is correlated with patch index.
}
\keyword{datasets}
