% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df1_phyl}
\alias{df1_phyl}
\title{Random phylogenetic tree for the simulated regional pool}
\format{
A \code{phylo}-class object with 450 tips, 449 nodes.
}
\source{
\code{library(ape)}

\code{df1_phyl <- ape::rtree(n = 450)}
}
\usage{
df1_phyl
}
\description{
The results of a simulation of environmental filtering using the \code{FilterABM} package.
A random phylogeny generated with the \code{ape} package.
}
\keyword{datasets}
