% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df1_trait_cat_dist}
\alias{df1_trait_cat_dist}
\title{Assumed distance matrix for categories of a categorical trait.}
\format{
A 5x5 matrix.
}
\source{
\code{df1_trait_cat_dist <- outer(c(-4.5, -3, -1, 1.5, 3.5), c(-4.5, -3, -1, 1.5, 3.5), '-') \%>\% abs()}
}
\usage{
df1_trait_cat_dist
}
\description{
In the \code{df1_com} dataset, the categorical trait was generated as a K-means (K=5) clustering of the continuous trait.
This matrix describes dissimilarity between clusters.
}
\keyword{datasets}
