% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_funcs.R
\name{dist_cat}
\alias{dist_cat}
\title{Distance function for a categorical trait}
\usage{
dist_cat(df1, df2, trait_idx, abun_idx = NULL, dmatrix)
}
\arguments{
\item{df1}{A dataset describing community one with at least one column corresponding to trait values}

\item{df2}{A dataset describing community two with at least one column corresponding to trait values; \code{df1} and \code{df2} may have different length}

\item{trait_idx}{Character; column name of the column with trait values}

\item{abun_idx}{Character; column name of the column with species abundance values; leave \code{NULL} if rows in \code{df1} and \code{df2} represent individuals but not species; if not \code{NULL}, corresponding columns in \code{df1} and \code{df2} must be integer}

\item{dmatrix}{Square matrix; distance matrix between the values of the categorical trait; make sure that col/row-names are the same as all possible trait values}
}
\value{
A mean distance (as specified in distance matrix) between all individual traits across the two communities.
}
\description{
Distance function for a categorical trait
}
\examples{
df1 = data.frame(trait = c(1, 2, 3, 1, 2, 3),
  n = rpois(n = 6, lambda = 5))
df2 = data.frame(trait = c(1, 1, 2, 3, 4),
  n = rpois(n = 5, lambda = 5))
dmatrix = matrix(c(0, 1, 2, 3, 1, 0, 1, 2, 2, 1, 0, 1, 3, 2, 1, 0),
  nrow = 4, byrow = TRUE)
colnames(dmatrix) = 1:4; rownames(dmatrix) = 1:4
dist_cat(df1, df2, trait_idx = "trait", abun_idx = "n",
  dmatrix = dmatrix)
}
