% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_funcs.R
\name{dist_cont}
\alias{dist_cont}
\title{Distance function for a continuous trait}
\usage{
dist_cont(df1, df2, trait_idx, abun_idx = NULL)
}
\arguments{
\item{df1}{A dataset describing community one with at least one column corresponding to trait values}

\item{df2}{A dataset describing community two with at least one column corresponding to trait values; \code{df1} and \code{df2} may have different length}

\item{trait_idx}{Character; column name of the column with trait values}

\item{abun_idx}{Character; column name of the column with species abundance values; leave \code{NULL} if rows in \code{df1} and \code{df2} represent individuals but not species}
}
\value{
A mean absolute difference between all individual traits across communities.
}
\description{
Distance function for a continuous trait
}
\examples{
df1 = data.frame(trait = rnorm(n = 3, mean = -1),
  n = rpois(n = 3, lambda = 5))
df2 = data.frame(trait = rnorm(n = 5, mean = 1),
  n = rpois(n = 5, lambda = 5))
dist_cont(df1, df2, trait_idx = "trait", abun_idx = "n")
}
