% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_funcs.R
\name{dist_distr}
\alias{dist_distr}
\title{Distance function for a distributional trait}
\usage{
dist_distr(df1, df2, trait_idx, abun_idx = NULL)
}
\arguments{
\item{df1}{A dataset describing community one with at least one column corresponding to trait values}

\item{df2}{A dataset describing community two with at least one column corresponding to trait values; \code{df1} and \code{df2} may have different length}

\item{trait_idx}{Character; column name of the column with trait values}

\item{abun_idx}{Character; column name of the column with species abundance values; leave \code{NULL} if rows in \code{df1} and \code{df2} represent individuals but not species}
}
\value{
A mean sum of deviations across discrete distributions across all individuals in the two communities.
}
\description{
Distance function for a distributional trait
}
\examples{
df1 = data.frame(n = rpois(n = 6, lambda = 5),
  t1 = c(1, 1, 1, 0.5, 0.5, 0.25),
    t2 = 1 - c(1, 1, 1, 0.5, 0.5, 0.25))
df2 = data.frame(n = rpois(n = 5, lambda = 5),
  t1 = c(0.1, 0.2, 0.5, 0.7, 0.8),
    t2 = 1 - c(0.1, 0.2, 0.5, 0.7, 0.8))
dist_distr(df1, df2, trait_idx = c("t1", "t2"), abun_idx = "n")
}
