% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_funcs.R
\name{dist_phyl}
\alias{dist_phyl}
\title{Distance function for community phylogenies}
\usage{
dist_phyl(df1, df2, taxa_idx, abun_idx = NULL, phtree)
}
\arguments{
\item{df1}{A dataset describing community one with at least one column corresponding to trait values}

\item{df2}{A dataset describing community two with at least one column corresponding to trait values; \code{df1} and \code{df2} may have different length}

\item{taxa_idx}{Character; column name of the column with taxa identities}

\item{abun_idx}{Character; column name of the column with species abundance values; leave \code{NULL} if rows in \code{df1} and \code{df2} represent individuals but not species; if not \code{NULL}, corresponding columns in \code{df1} and \code{df2} must be integer}

\item{phtree}{Phylogenetic tree (\code{ape} class "\code{phylo}") that includes all taxa from both communities}
}
\value{
A phylogenetic distance between two communities.
}
\description{
Distance function for community phylogenies
}
\examples{
df1 = data.frame(species = paste0("t", 1:6),
  n = rpois(n = 6, lambda = 5))
df2 = data.frame(species = paste0("t", 3:7),
  n = rpois(n = 5, lambda = 5))
phtree <- ape::rtree(10)
dist_phyl(df1, df2, taxa_idx = "species", abun_idx = "n",
  phtree = phtree)
}
