% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eftest.R
\name{eftest}
\alias{eftest}
\title{Generate AIC-weighted SCAM spline mean derivative for the effect of an environmental variable on community trait structure given spatial and phylogenetic constraints}
\usage{
eftest(
  comdata,
  envdata,
  npairs = 1000,
  rep = 1,
  env_idx,
  com_idx,
  trait_idx,
  taxa_idx,
  abun_idx = NULL,
  x_idx,
  y_idx,
  trait_FUN,
  phyl_FUN,
  phtree,
  ...
)
}
\arguments{
\item{comdata}{Data frame containing the data on structure of communities: it should include community identifiers, species identities, trait values, and (optionally) species abundances.}

\item{envdata}{Data frame containing the data on environmental conditions: it should include community identifiers (values equivalent to those in \code{comdata}), environmental factor levels, and geographical coordinates (for large scale, make sure to transform the coordinate system).}

\item{npairs}{Integer. Number of random community pairs to pick; this function will not pick more than a square of number of communities in the data (e.g., all pairs will be unique).}

\item{rep}{Positive integer, how many times to repeat the procedure (set a larger number to estimate uncertainity, but it is computationally expensive)}

\item{env_idx}{String, the name of a column in \code{envdata} that contains the values of the \strong{environmental factor}.}

\item{com_idx}{String, the name of a column in \code{envdata} and \code{comdata} that contains \strong{community identifiers}. The column name should be the same in both data frames.}

\item{trait_idx}{String, the name of a column in \code{comdata} that contains the data on \strong{trait values}.}

\item{taxa_idx}{String, the name of a column in \code{comdata} that contains \strong{species identifiers}.}

\item{abun_idx}{\code{NULL} or string, the name of a column in \code{comdata} that contains \strong{species counts}. If \code{NULL}, all counts are assumed to be equal to one (e.g., when rows in \code{comdata} correspond to individuals rather than species).}

\item{x_idx}{String, the name of a column in \code{envdata} that contain the \strong{x-coordinate} of a community (e.g., longitude).}

\item{y_idx}{String, the name of a column in \code{envdata} that contain the \strong{y-coordinate} of a community (e.g., latitude).}

\item{trait_FUN}{Function, a distance function to assess trait-based distance between the communities. Options include \code{dist_cont}, \code{dist_cat}, \code{dist_distr}, or a manually written function.}

\item{phyl_FUN}{Function, a distance function to assess phylogeny-based distance between the communities. Options include \code{dist_phyl} or a manually written function. If missing but \code{phtree} is not, the \code{enfiltest::dist_phyl} is used by default. If both are missing, phylogenetic distance is not estimated.}

\item{phtree}{Phylogenetic tree (\code{ape} class "\code{phylo}") that includes all taxa from all communities in \code{comdata}. If missing but \code{phyl_FUN} is not, a random tree is generated. If both are missing, phylogenetic distance is not estimated.}

\item{...}{Additional arguments for \code{trait_FUN}.}
}
\value{
A vector of length \code{rep} with AIC-weighted SCAM spline derivative means for each run.
}
\description{
Generate AIC-weighted SCAM spline mean derivative for the effect of an environmental variable on community trait structure given spatial and phylogenetic constraints
}
