% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eftest.R
\name{eftest_gam_dx}
\alias{eftest_gam_dx}
\title{Return a mean SCAM spline derivative}
\usage{
eftest_gam_dx(fit)
}
\arguments{
\item{fit}{A \code{scam::scam} model as constructed by \code{eftest_models()[[2]]}}
}
\value{
A mean derivative
}
\description{
Return a mean SCAM spline derivative
}
