% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eftest.R
\name{eftest_models}
\alias{eftest_models}
\title{Return a set of GAMs explaining difference in community trait structure}
\usage{
eftest_models(rpd)
}
\arguments{
\item{rpd}{A tibble, output of the \code{rand_pair_diff} function containing columns for spatial distance between communities (\code{d_dist}), difference in an environmental level (\code{d_env}), difference in trait structure (\code{d_trait}), and (optional) difference in phylogenetic structure (\code{d_phyl}).}
}
\value{
A list of lists with two elements each: a vector indicating whether a spatial/environmental/phylogenetic distance term was included in the model and the \code{scam::scam} model object.
}
\description{
Return a set of GAMs explaining difference in community trait structure
}
\details{
The null model is assumed to be a simple intercept model.

For the term of spatial distance between community locations, a simple cubic regression spline function is used.

For the terms corresponding to difference in environmental level and phylogenetic structure, a shape-constrained additive model is used with a monotonically increasing spline function.
}
\examples{
test_comdata = df1_com \%>\% dplyr::mutate(n = 1)
test_envdata <- df1_env \%>\% dplyr::mutate(x = patch, y = 0)
x = rand_pair_diff(comdata = test_comdata,
  npairs = 100,
  envdata = test_envdata,
  env_idx = "env",
  com_idx = "patch",
  trait_idx = "trait_cont",
  taxa_idx = "species",
  abun_idx = NULL,
  x_idx = "x", y_idx = "y",
  trait_FUN = dist_cont, phyl_FUN = dist_phyl,
  phtree = df1_phyl)
eftest_models(x)
}
