% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eftest.R
\name{eftest_rankAIC}
\alias{eftest_rankAIC}
\title{Generate AIC table with a mean SCAM spline derivative}
\usage{
eftest_rankAIC(models)
}
\arguments{
\item{models}{A list of \code{scam::scam} models as constructed by \code{eftest_models()}}
}
\value{
A tibble with columns indicating terms in the models, AIC and ΔAIC, log-likelihood, estimated degrees of freedom, AIC weight, and a mean derivative for \code{d_env} SCAM term.
}
\description{
Generate AIC table with a mean SCAM spline derivative
}
