% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eftest.R
\name{eftest_stat}
\alias{eftest_stat}
\title{Get AIC-weighted SCAM spline derivative for env term}
\usage{
eftest_stat(efAICtable)
}
\arguments{
\item{efAICtable}{An AIC table generated by \code{eftest_rankAIC()}}
}
\value{
AIC-weighted SCAM spline derivative for env term
}
\description{
Get AIC-weighted SCAM spline derivative for env term
}
