% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{\%notin\%}
\alias{\%notin\%}
\title{A "not in" operator}
\usage{
x \%notin\% y
}
\arguments{
\item{x}{vector or NULL: the values to be matched}

\item{y}{vector or NULL: the values to be matched against}
}
\value{
An opposite of the \code{match} operator.
}
\description{
A "not in" operator
}
