% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eftest.R
\name{rand_pair_diff}
\alias{rand_pair_diff}
\title{Pairwise differences between random communities}
\usage{
rand_pair_diff(
  comdata,
  envdata,
  npairs = 1000,
  env_idx,
  com_idx,
  trait_idx,
  taxa_idx,
  abun_idx = NULL,
  x_idx,
  y_idx,
  trait_FUN,
  phyl_FUN,
  phtree,
  ...
)
}
\arguments{
\item{comdata}{Data frame containing the data on structure of communities: it should include community identifiers, species identities, trait values, and (optionally) species abundances.}

\item{envdata}{Data frame containing the data on environmental conditions: it should include community identifiers (values equivalent to those in \code{comdata}), environmental factor levels, and geographical coordinates (for large scale, make sure to transform the coordinate system).}

\item{npairs}{Integer. Number of random community pairs to pick; this function will not pick more than a square of number of communities in the data (e.g., all pairs will be unique).}

\item{env_idx}{String, the name of a column in \code{envdata} that contains the values of the \strong{environmental factor}.}

\item{com_idx}{String, the name of a column in \code{envdata} and \code{comdata} that contains \strong{community identifiers}. The column name should be the same in both data frames.}

\item{trait_idx}{String, the name of a column in \code{comdata} that contains the data on \strong{trait values}.}

\item{taxa_idx}{String, the name of a column in \code{comdata} that contains \strong{species identifiers}.}

\item{abun_idx}{\code{NULL} or string, the name of a column in \code{comdata} that contains \strong{species counts}. If \code{NULL}, all counts are assumed to be equal to one (e.g., when rows in \code{comdata} correspond to individuals rather than species).}

\item{x_idx}{String, the name of a column in \code{envdata} that contain the \strong{x-coordinate} of a community (e.g., longitude).}

\item{y_idx}{String, the name of a column in \code{envdata} that contain the \strong{y-coordinate} of a community (e.g., latitude).}

\item{trait_FUN}{Function, a distance function to assess trait-based distance between the communities. Options include \code{dist_cont}, \code{dist_cat}, \code{dist_distr}, or a manually written function.}

\item{phyl_FUN}{Function, a distance function to assess phylogeny-based distance between the communities. Options include \code{dist_phyl} or a manually written function. If missing but \code{phtree} is not, the \code{enfiltest::dist_phyl} is used by default. If both are missing, phylogenetic distance is not estimated.}

\item{phtree}{Phylogenetic tree (\code{ape} class "\code{phylo}") that includes all taxa from all communities in \code{comdata}. If missing but \code{phyl_FUN} is not, a random tree is generated. If both are missing, phylogenetic distance is not estimated.}

\item{...}{Additional arguments for \code{trait_FUN}.}
}
\value{
A tibble with identifiers of communities in a pair (\code{c1, c2}), topological (Euclidean) distance between the communities (\code{d_dist}), difference in environmental conditions (\code{d_env}), phylogenetic difference (\code{d_phyl}), and trait difference (\code{d_trait}).
}
\description{
Estimating difference in topological location, environmental conditions, trait structure, and (if phylogeny provided) phylogenetic relatedness between random pairs of communities.
}
\examples{
test_comdata = df1_com \%>\% dplyr::mutate(n = 1)
test_envdata <- df1_env \%>\% dplyr::mutate(x = patch, y = 0)
rand_pair_diff(comdata = test_comdata,
  npairs = 10,
  envdata = test_envdata,
  env_idx = "env",
  com_idx = "patch",
  trait_idx = "trait_cont",
  taxa_idx = "species",
  abun_idx = NULL,
  x_idx = "x", y_idx = "y",
  trait_FUN = dist_cont, phyl_FUN = dist_phyl,
  phtree = df1_phyl)
}
