% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rao_distances.R
\name{tree2phylmatr}
\alias{tree2phylmatr}
\title{Get phylogenetic distances matrix from a phylogenetic tree}
\usage{
tree2phylmatr(phtree)
}
\arguments{
\item{phtree}{Phylogenetic tree, a 'phylo' class object (see 'ape' package)}
}
\value{
Phylogenetic distance matrix
}
\description{
A formal wrapper around the original \code{ape::cophenetic.phylo()} function.
}
\examples{
ape::rtree(n = 5) \%>\% tree2phylmatr()

}
