% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{ucon}
\alias{ucon}
\title{Pair sampling without edge effects}
\usage{
ucon(vec)
}
\arguments{
\item{vec}{A vector}
}
\value{
A pair of values
}
\description{
Pairwise sampling of a sample of values drawn from a uniform distribution will inadvertedly produce such a sample that an absolute difference across the value pairs will have a triangular distribution.
E.g., \code{x <- runif(10000); y <- sample(x, 10000) - sample(x, 10000); hist(abs(y))}.

This function attempts to balance pairwise sampling in such a way that the differences across value pairs are distributed normally.
Note that the output is still skewed but not as triangular.
}
\examples{
x <- runif(10000)
y <- sapply(1:1000, function(i) ucon(x))
hist(abs(y[1,] - y[2,]))
}
