% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eftest.R
\name{eftest_stat}
\alias{eftest_stat}
\title{Get AIC-weighted SCAM spline derivative for env term}
\usage{
eftest_stat(efAICtable, aic_threshold = 2)
}
\arguments{
\item{efAICtable}{An AIC table generated by \code{eftest_rankAIC()}}

\item{aic_threshold}{Threshold of delta-AIC for the null model; if the null model is within this value of delta-AIC, the test score is automatically zero; default to delta-AIC = 2.}
}
\value{
AIC-weighted SCAM spline derivative for env term
}
\description{
Get AIC-weighted SCAM spline derivative for env term
}
